updated verbiage
Deleted Principle (Flinto) and Google Docs from "Materials" because there was no mention of animation in the rest of the brief. 
I actually added some additional wording rather than deleting. I think for this particular project, it doesn't hurt to have the added information, as it provides a much clearer roadmap. I think the more clarity that is offered on expectations, the less they will overthink it. If I were to remove one section, it would probably be "Design Thinking" as it is a bit repetitive.